+++ CHANGELOG.md
@@ -1,16 +1,28 @@
 # Change Log
 
-ver 0.3.1
+ver 0.4
 
+
+### 0.4
+
+    add:
+
+    + pack-unpack data section
+
+    + use dirs category on unpack
+
+    remove:
+
+    + 'prefix' flag
 
 ### 0.3
 
-+ unpack mode from export JSON file to files
+    + unpack mode from export JSON file to files
 
 ### 0.2
 
-+ pack menus, blocks, langs, tables, params section from files to export JSON file
+    + pack menus, blocks, langs, tables, params section from files to export JSON file
 
 ### 0.1
 
-+ pack contracts and pages section from files to export JSON file
+    + pack contracts and pages section from files to export JSON file